refactor : 필터 검색 시 페이지네이션을 프론트단에서 처리

+++ server/src/main/java/project/server/domain/cocktail/embed/rate/Rate.java
@@ -2,6 +2,7 @@ package project.server.domain.cocktail.embed.rate;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
 import javax.persistence.Embeddable;
 
 @Embeddable
@@ -11,18 +12,18 @@ public class Rate {
 
     private double rate = 0;
 
-    private double ratedCount = 0;
+    private long ratedCount = 0;
 
-    private double totalRateSum = 0;
+    private long totalRateSum = 0;
 
     public void calculate(int value) {
         ratedCount++;
         totalRateSum += value;
-        rate = Math.round(totalRateSum/ratedCount*100)/100.0;
+        rate = Math.round((double) totalRateSum / ratedCount * 100) / 100.0;
     }
 
     public void reCalculate(int oldValue, int value) {
         totalRateSum += value - oldValue;
-        rate = Math.round(totalRateSum/ratedCount*100)/100.0;
+        rate = Math.round((double) totalRateSum / ratedCount * 100) / 100.0;
     }
 }
